requirements.txt
# AgentProject — 版本对齐自当前 .venv（pip freeze）
# 安装：pip install -r requirements.txt
#
# 如需完整可复现快照（含全部传递依赖）：pip freeze > requirements-lock.txt

# ---------- Web API（main.py）----------
# fastapi 若 pip 提示无 wheel，请换 Python 3.11/3.12 或稍后重试
fastapi>=0.115.0
uvicorn[standard]==0.42.0

# ---------- Web UI（app.py）----------
streamlit==1.55.0

# ---------- LangChain / Agent / LangGraph ----------
langchain==1.2.12
langchain-core==1.2.19
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.20

# ---------- 向量库 / RAG ----------
chromadb==1.5.5

# ---------- 阿里云 DashScope（通义千问 / Embedding）----------
dashscope==1.25.14

# ---------- 通用工具 ----------
requests==2.32.5
PyYAML==6.0.3
pypdf==6.9.1
python-dotenv==1.2.2
httpx==0.28.1

# ---------- 数据与校验（生态常用）----------
pydantic==2.12.5
pydantic-settings==2.13.1
numpy==2.4.3

# ---------- Vue Frontend（frontend/）----------
# 前端请在 frontend/ 目录执行 npm install && npm run dev

# ---------- 生产级数据库与任务队列 ----------
sqlalchemy==2.1.4
psycopg2-binary==2.9.13
celery==5.4.0
redis==5.2.1
APScheduler==3.11.1
prometheus-client==0.21.0
